smart-ml-forecaster/requirements.txt
kafka-python
pandas
scikit-learn